+++ script/build-configuration.js
@@ -7,7 +7,7 @@ const fs = require('node:fs');
 const path = require('node:path');
 
 console.log(`Running ${__filename} ...`);
-console.log('Setting build configuration...');
+console.warn('Setting build configuration...');
 const package = JSON.parse(fs.readFileSync(path.resolve(__dirname, '..', 'package.json')).toString());
 const debug = Boolean(process.env.DEBUG ?? false);
 const buildConfiguration = {
@@ -15,5 +15,5 @@ const buildConfiguration = {
     debug,
 }
 fs.writeFileSync(path.resolve(__dirname, '..', 'src', 'build-configuration.json'), JSON.stringify(buildConfiguration, null, 2));
-console.log('Setting build configuration done.');
+console.warn('Setting build configuration done.');
 console.log(`Running ${__filename} done.`);